third_party/blink/web_tests/fast/ruby/floating-ruby-text-expected.txt

CONSOLE MESSAGE: `float` property is not supported correctly inside an element with `display: ruby` or `display: ruby-text`.
